move NativeTypeInstance to own file
