+++ libs/datamodel/connectors/datamodel-connector/src/lib.rs
@@ -4,7 +4,7 @@ pub mod error;
 use crate::error::ConnectorError;
 pub use combined_connector::CombinedConnector;
 use dml::native_type_constructor::NativeTypeConstructor;
-use dml::NativeTypeInstance;
+use dml::native_type_instance::NativeTypeInstance;
 
 pub trait Connector: Send + Sync {
     fn capabilities(&self) -> &Vec<ConnectorCapability>;
