+++ libs/datamodel/core/src/dml/mod.rs
@@ -3,8 +3,8 @@ pub use dml::default_value::*;
 pub use dml::field::*;
 pub use dml::model::*;
 pub use dml::native_type_constructor::*;
+pub use dml::native_type_instance::*;
 pub use dml::r#enum::*;
 pub use dml::relation_info::*;
 pub use dml::scalars::*;
 pub use dml::traits::*;
-pub use dml::NativeTypeInstance;
